Move SpanKey to internal package
