+++ instrumentation/opentelemetry-api/opentelemetry-api-1.0/testing/src/main/java/AgentSpanTesting.java
@@ -3,6 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import io.opentelemetry.instrumentation.api.internal.SpanKey;
+
 public class AgentSpanTesting {
 
   /**
@@ -23,8 +25,7 @@ public class AgentSpanTesting {
 
   /**
    * Runs the provided {@code runnable} inside the scope of an INTERNAL span with name {@code
-   * spanName}. Span is added into context under all possible keys from {@link
-   * io.opentelemetry.instrumentation.api.instrumenter.SpanKey}
+   * spanName}. Span is added into context under all possible keys from {@link SpanKey}
    */
   public static void runWithAllSpanKeys(String spanName, Runnable runnable) {
     runnable.run();
